portfolio-aggregator/app/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = ""
authors = [
    {name = "ae",email = "alex@gzcim.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry]
name = "back-end"
version = "0.1.0"
packages = [{ include = "app" }]
